requirements.txt
Django==4.2
djangorestframework==3.15.0
requests==2.31.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
